requirements.txt
gunicorn==20.0.4
WebOb==1.8.6
parse==1.18.0
pytest==6.2.1
requests==2.25.1
requests-wsgi-adapter==0.4.1
pytest-cov==2.10.1
jinja2==2.11.2

